pull of last changes

+++ src/main/java/pl/schoolmanager/controller/DivisionController.java
@@ -1,11 +1,15 @@
 package pl.schoolmanager.controller;
 
 import java.util.List;
+import java.util.ListIterator;
 
+import javax.persistence.PersistenceException;
+import javax.servlet.http.HttpSession;
+import javax.validation.ConstraintViolationException;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.orm.jpa.JpaSystemException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -15,17 +19,17 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import pl.schoolmanager.bean.SessionManager;
 import pl.schoolmanager.entity.Division;
 import pl.schoolmanager.entity.Mark;
+import pl.schoolmanager.entity.School;
 import pl.schoolmanager.entity.Student;
 import pl.schoolmanager.entity.Subject;
-import pl.schoolmanager.entity.User;
 import pl.schoolmanager.repository.DivisionRepository;
 import pl.schoolmanager.repository.MarkRepository;
 import pl.schoolmanager.repository.MessageRepository;
 import pl.schoolmanager.repository.StudentRepository;
 import pl.schoolmanager.repository.SubjectRepository;
-import pl.schoolmanager.repository.UserRepository;
 
 @Controller
 @RequestMapping("/division")
@@ -42,19 +46,18 @@ public class DivisionController {
 
 	@Autowired
 	private MarkRepository markRepository;
-	
-	@Autowired
-	private UserRepository userRepository;
 
 	@Autowired
 	private MessageRepository messageRepository;
+
+	@Autowired
+	private SessionManager sessionManager;
 
 	@GetMapping("/all")
 	public String all(Model m) {
 		return "division/all_divisions";
 	}
 
-	// CREATE
 	@GetMapping("/create")
 	public String createDivision(Model m) {
 		m.addAttribute("division", new Division());
@@ -70,7 +73,6 @@ public class DivisionController {
 		return "index";
 	}
 
-	// READ
 	@GetMapping("/view/{divisionId}")
 	public String viewDivision(Model m, @PathVariable long divisionId) {
 		Division division = this.divisionRepository.findOne(divisionId);
@@ -78,7 +80,6 @@ public class DivisionController {
 		return "division/show_division";
 	}
 
-	// UPDATE
 	@GetMapping("/update/{divisionId}")
 	public String updateDivision(Model m, @PathVariable long divisionId) {
 		Division division = this.divisionRepository.findOne(divisionId);
@@ -97,14 +98,34 @@ public class DivisionController {
 		return "index";
 	}
 
-	// DELETE
 	@GetMapping("/delete/{divisionId}")
-	public String deleteDivision(@PathVariable long divisionId) {
-		this.divisionRepository.delete(divisionId);
-		return "index";
+	public String deleteDivision(@PathVariable long divisionId, Model m) {
+		//Code to add if confirm button need to be implemented in all_divisions view
+//		Long schoolId = this.divisionRepository.findOne(divisionId).getId();
+//		List<Division> schoolDivisions = this.divisionRepository.findAllBySchoolId(schoolId);
+//		m.addAttribute("schoolDivisions", schoolDivisions);
+//		m.addAttribute("availableDivisions", this.divisionRepository.findAll());
+//		m.addAttribute("remove", divisionId);
+//		return "division/all_divisions";
+		Division division = this.divisionRepository.findOne(divisionId);
+		m.addAttribute("division", division);
+		return "division/confirmdelete_divisions";
+	}
+	
+	@PostMapping("/delete/{divisionId}")
+	public String deleteSchool(@PathVariable long divisionId) {
+		Division division = this.divisionRepository.findOne(divisionId);
+		if (division.getSubject()!=null || division.getStudent() != null || division.getSchool() != null) {
+			return "errors/deleteException";
+		}
+		try {
+			this.divisionRepository.delete(divisionId);
+		} catch (ConstraintViolationException | PersistenceException | JpaSystemException e) {
+			return "errors/deleteException";
+		}
+		return "redirect:/division/all";
 	}
 
-	// INSIDE DIVISION
 	@GetMapping("/inside/{divisionId}")
 	public String insideDivision(Model m, @PathVariable long divisionId) {
 		Division division = this.divisionRepository.findOne(divisionId);
@@ -116,7 +137,6 @@ public class DivisionController {
 		return "division/inside_division";
 	}
 
-	// ADD STUDENT TO DIVISION
 	@GetMapping("/addStudent/{divisionId}")
 	public String addStudent(Model m, @PathVariable long divisionId) {
 		Division division = this.divisionRepository.findOne(divisionId);
@@ -137,7 +157,15 @@ public class DivisionController {
 		return "redirect:/division/addStudent/{divisionId}";
 	}
 
-	// ADD SUBJECT TO DIVISION
+	@GetMapping("removeStudent/{divisionId}/{studentId}")
+	public String removeStudent(@PathVariable long divisionId, @PathVariable long studentId) {
+		Division division = this.divisionRepository.findOne(divisionId);
+		Student student = this.studentRepository.findOne(studentId);
+		student.setDivision(null);
+		this.studentRepository.save(student);
+		return "redirect:/division/addStudent/{divisionId}";
+	}
+
 	@GetMapping("/addSubject/{divisionId}")
 	public String addSubject(Model m, @PathVariable long divisionId) {
 		Division division = this.divisionRepository.findOne(divisionId);
@@ -159,7 +187,20 @@ public class DivisionController {
 		return "redirect:/division/addSubject/{divisionId}";
 	}
 
-	// ALL STUDENTS IN DIVISION
+	@GetMapping("removeSubject/{divisionId}/{subjectId}")
+	public String removeSubject(@PathVariable long divisionId, @PathVariable long subjectId) {
+		Division division = this.divisionRepository.findOne(divisionId);
+		Subject subject = this.subjectRepository.findOne(subjectId);
+		ListIterator<Division> div = subject.getDivision().listIterator();
+		while (div.hasNext()) {
+			if (div.next().getId() == division.getId()) {
+				div.remove();
+			}
+		}
+		this.subjectRepository.save(subject);
+		return "redirect:/division/addSubject/{divisionId}";
+	}
+
 	@GetMapping("/inside/students/{divisionId}")
 	public String studentsInsideDivision(Model m, @PathVariable long divisionId) {
 		Division division = this.divisionRepository.findOne(divisionId);
@@ -171,7 +212,6 @@ public class DivisionController {
 		return "division/allStudents_division";
 	}
 
-	// ALL SUBJECTS IN DIVISION
 	@GetMapping("/inside/subjects/{divisionId}")
 	public String subjectsInsideDivision(Model m, @PathVariable long divisionId) {
 		Division division = this.divisionRepository.findOne(divisionId);
@@ -181,7 +221,6 @@ public class DivisionController {
 		return "division/allSubjects_division";
 	}
 
-	// ALL MARKS IN SUBJECT IN DIVISION
 	@GetMapping("/inside/marks/{divisionId}/{subjectId}")
 	public String subjectsMarksInsideDivision(Model m, @PathVariable long divisionId, @PathVariable long subjectId) {
 		Division division = this.divisionRepository.findOne(divisionId);
@@ -197,31 +236,17 @@ public class DivisionController {
 		return "division/allStudentsMarks_division";
 	}
 
-	// SHOW ALL
+	@ModelAttribute("availableDivisions")
+	public List<Division> getSchoolDivisions() {
+		HttpSession s = SessionManager.session();
+		School school = (School) s.getAttribute("thisSchool");
+		return this.divisionRepository.findAllBySchool(school);
+	}
+
+	// SHOW ALL DIVISIONS IN SCHOOL
 	@ModelAttribute("availableDivisions")
 	public List<Division> getDivisions() {
 		return this.divisionRepository.findAll();
 	}
 
-	// MESSAGES INFO
-	@ModelAttribute("countAllReceivedMessages")
-	public Integer countAllReceivedMessages(Long receiverId) {
-		return this.messageRepository.findAllByReceiverId(getLoggedUser().getId()).size();
-	}
-
-	@ModelAttribute("countAllSendedMessages")
-	public Integer countAllSendedMessages(Long senderId) {
-		return this.messageRepository.findAllBySenderId(getLoggedUser().getId()).size();
-	}
-	
-	@ModelAttribute("countAllReceivedUnreadedMessages")
-	public Integer countAllReceivedUnreadedMessages(Long receiverId, Integer checked) {
-		return this.messageRepository.findAllByReceiverIdAndChecked(getLoggedUser().getId(), 0).size();
-	}
-	
-	private User getLoggedUser() {
-		Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-		String username = ((org.springframework.security.core.userdetails.User) principal).getUsername();
-		return this.userRepository.findOneByUsername(username);
-	}
 }
