+++ src/main/java/pl/schoolmanager/controller/StudentController.java
@@ -1,32 +1,25 @@
 package pl.schoolmanager.controller;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.servlet.http.HttpSession;
-import javax.validation.Valid;
-
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
+import org.springframework.web.bind.annotation.*;
 import pl.schoolmanager.bean.SessionManager;
 import pl.schoolmanager.entity.School;
 import pl.schoolmanager.entity.Student;
+import pl.schoolmanager.entity.Subject;
+import pl.schoolmanager.entity.Teacher;
 import pl.schoolmanager.entity.User;
 import pl.schoolmanager.entity.UserRole;
 import pl.schoolmanager.repository.MessageRepository;
 import pl.schoolmanager.repository.SchoolRepository;
 import pl.schoolmanager.repository.StudentRepository;
-import pl.schoolmanager.repository.UserRepository;
 
+import javax.servlet.http.HttpSession;
+import javax.validation.Valid;
+import java.util.ArrayList;
+import java.util.List;
 
 @Controller
 @RequestMapping("/student")
@@ -34,15 +27,16 @@ public class StudentController {
 
 	@Autowired
 	private StudentRepository studentRepository;
-	@Autowired
-	private UserRepository userRepo;
+
 	@Autowired
 	private SchoolRepository schoolRepo;
 
 	@Autowired
 	private MessageRepository messageRepository;
-	
-	// CREATE
+
+	@Autowired
+	private SessionManager sessionManager;
+
 	@GetMapping("/create")
 	public String createStudent(Model m) {
 		m.addAttribute("student", new Student());
@@ -62,48 +56,56 @@ public class StudentController {
 	@GetMapping("/userNewStudent")
 	public String newStudentFromUser(Model m) {
 		m.addAttribute("student", new Student());
+		HttpSession s = SessionManager.session();
+		s.setAttribute("thisSchoolAdmin", null);
+		s.setAttribute("thisTeacher", null);
+		s.setAttribute("thisStudent", null);
 		return "student/user_new_student";
 	}
 
 	@PostMapping("/userNewStudent")
-	public String newStudentFromUserPost(@Valid @ModelAttribute Student student, BindingResult bindingResult,
-										Model m) {
+	public String newStudentFromUserPost(@Valid @ModelAttribute Student student, BindingResult bindingResult, Model m) {
 		if (bindingResult.hasErrors()) {
 			return "student/user_new_student";
 		}
-		User user = getLoggedUser();
+		User user = sessionManager.loggedUser();
 		UserRole userRole = new UserRole();
 		userRole.setUsername(user.getUsername());
 		userRole.setUserRole("ROLE_STUDENT");
 		userRole.setSchool(student.getSchool());
 		userRole.setUser(user);
+		userRole.setEnabled(false);
 		student.setUserRole(userRole);
 		this.studentRepository.save(student);
 		HttpSession s = SessionManager.session();
 		s.setAttribute("thisStudent", student);
 		s.setAttribute("thisSchool", student.getSchool());
-		return "redirect:/studentView/division";
+		return "user/user_activation_info";
 	}
-	
+
 	// Managing exisitng student role
 	@GetMapping("/userStudent")
 	public String StudentFromUser(Model m) {
 		m.addAttribute("student", new Student());
+		HttpSession s = SessionManager.session();
+		s.setAttribute("thisSchoolAdmin", null);
+		s.setAttribute("thisTeacher", null);
+		s.setAttribute("thisStudent", null);
 		return "student/user_student";
 	}
 
 	@PostMapping("/userStudent")
-	public String StudentFromUserPost(@Valid @ModelAttribute Student student, BindingResult bindingResult,
-										Model m) {
+	public String StudentFromUserPost(@Valid @ModelAttribute Student student, BindingResult bindingResult, Model m) {
 		if (bindingResult.hasErrors()) {
 			return "student/user_student";
 		}
-		User user = getLoggedUser();
+		User user = sessionManager.loggedUser();
 		Student thisStudent = null;
 		UserRole thisUserRole = null;
 		List<UserRole> userRoles = user.getUserRoles();
 		for (UserRole userRole : userRoles) {
-			if (userRole.getUserRole().equals("ROLE_STUDENT") && userRole.getSchool().getName().equals(student.getSchool().getName())) {
+			if (userRole.getUserRole().equals("ROLE_STUDENT")
+					&& userRole.getSchool().getName().equals(student.getSchool().getName())) {
 				thisUserRole = userRole;
 			}
 		}
@@ -116,10 +118,9 @@ public class StudentController {
 		HttpSession s = SessionManager.session();
 		s.setAttribute("thisStudent", thisStudent);
 		s.setAttribute("thisSchool", student.getSchool());
-		return "redirect:/studentView/division";
+		return "redirect:/studentView/";
 	}
 
-	// READ
 	@GetMapping("/view/{studentId}")
 	public String viewStudent(Model m, @PathVariable long studentId) {
 		Student student = this.studentRepository.findOne(studentId);
@@ -127,7 +128,6 @@ public class StudentController {
 		return "student/show_student";
 	}
 
-	// UPDATE
 	@GetMapping("/update/{studentId}")
 	public String updateStudent(Model m, @PathVariable long studentId) {
 		Student student = this.studentRepository.findOne(studentId);
@@ -146,29 +146,39 @@ public class StudentController {
 		return "index";
 	}
 
-	// DELETE
 	@GetMapping("/delete/{studentId}")
-	public String deleteStudent(@PathVariable long studentId) {
-		this.studentRepository.delete(studentId);
-		return "index";
+	public String deleteStudent(@PathVariable long studentId, Model m) {
+		Student student = this.studentRepository.findOne(studentId);
+		m.addAttribute("student", student);
+		return "student/confirmdelete_student";
 	}
 
-	// SHOW ALL
+	@PostMapping("/delete/{studentId}")
+	public String deleteStudent(@PathVariable long studentId) {
+		Student student = this.studentRepository.findOne(studentId);
+		if (student.getSchool() != null || student.getMark() != null || student.getDivision() != null) {
+			return "errors/deleteException";
+		}
+		this.studentRepository.delete(studentId);
+		return "redirect:/student/all";
+	}
+
 	@ModelAttribute("availableStudents")
 	public List<Student> getStudents() {
 		return this.studentRepository.findAll();
 	}
 
+	// SHOW ALL FROM SCHOOL
+	@ModelAttribute("schoolStudents")
+	public List<Student> getSchoolStudents() {
+		HttpSession s = SessionManager.session();
+		School school = (School) s.getAttribute("thisSchool");
+		return this.studentRepository.findAllBySchool(school);
+	}
+
 	@GetMapping("/all")
 	public String all(Model m) {
 		return "student/all_students";
-	}
-
-	// Additional methods
-	private User getLoggedUser() {
-		Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-		String username = ((org.springframework.security.core.userdetails.User) principal).getUsername();
-		return this.userRepo.findOneByUsername(username);
 	}
 
 	@ModelAttribute("userRolesForSelect")
@@ -182,36 +192,32 @@ public class StudentController {
 		List<School> availableSchools = this.schoolRepo.findAll();
 		return availableSchools;
 	}
-	
 
-	@ModelAttribute("userSchools")
-	public List<School> userSchools() {
-		User user = getLoggedUser();
+	// Consider deleting this code
+	// @ModelAttribute("userSchools")
+	// public List<School> userSchools() {
+	// User user = sessionManager.loggedUser();
+	// List<School> schools = new ArrayList<>();
+	// List<UserRole> roles = user.getUserRoles();
+	// for (UserRole userRole : roles) {
+	// if (userRole.getUserRole().equals("ROLE_STUDENT")) {
+	// schools.add(userRole.getSchool());
+	// }
+	// }
+	// return schools;
+	// }
+
+	@ModelAttribute("userEnabledSchools")
+	public List<School> userEnabledSchools() {
+		User user = sessionManager.loggedUser();
 		List<School> schools = new ArrayList<>();
 		List<UserRole> roles = user.getUserRoles();
 		for (UserRole userRole : roles) {
-			if (userRole.getUserRole().equals("ROLE_STUDENT")) {
+			if (userRole.getUserRole().equals("ROLE_STUDENT") && userRole.isEnabled() == true) {
 				schools.add(userRole.getSchool());
 			}
 		}
 		return schools;
 	}
 
-
-	// MESSAGES INFO
-	@ModelAttribute("countAllReceivedMessages")
-	public Integer countAllReceivedMessages(Long receiverId) {
-		return this.messageRepository.findAllByReceiverId(getLoggedUser().getId()).size();
-	}
-
-	@ModelAttribute("countAllSendedMessages")
-	public Integer countAllSendedMessages(Long senderId) {
-		return this.messageRepository.findAllBySenderId(getLoggedUser().getId()).size();
-	}
-	
-	@ModelAttribute("countAllReceivedUnreadedMessages")
-	public Integer countAllReceivedUnreadedMessages(Long receiverId, Integer checked) {
-		return this.messageRepository.findAllByReceiverIdAndChecked(getLoggedUser().getId(), 0).size();
-	}
-	
 }
