+++ src/main/java/pl/schoolmanager/controller/SubjectController.java
@@ -5,7 +5,6 @@ import java.util.List;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -15,11 +14,10 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import pl.schoolmanager.bean.SessionManager;
 import pl.schoolmanager.entity.Subject;
-import pl.schoolmanager.entity.User;
 import pl.schoolmanager.repository.MessageRepository;
 import pl.schoolmanager.repository.SubjectRepository;
-import pl.schoolmanager.repository.UserRepository;
 
 @Controller
 @RequestMapping("/subject")
@@ -27,19 +25,18 @@ public class SubjectController {
 
 	@Autowired
 	private SubjectRepository subjectRepository;
-	
+
+	@Autowired
+	private MessageRepository messageRepository;
+
+	@Autowired
+	private SessionManager sessionManager;
+
 	@GetMapping("/all")
 	public String all(Model m) {
 		return "subject/all_subjects";
 	}
-	
-	@Autowired
-	private UserRepository userRepository;
 
-	@Autowired
-	private MessageRepository messageRepository;
-	
-	// CREATE
 	@GetMapping("/create")
 	public String createSubject(Model m) {
 		m.addAttribute("subject", new Subject());
@@ -55,7 +52,6 @@ public class SubjectController {
 		return "index";
 	}
 
-	// READ
 	@GetMapping("/view/{subjectId}")
 	public String viewSubject(Model m, @PathVariable long subjectId) {
 		Subject subject = this.subjectRepository.findOne(subjectId);
@@ -63,7 +59,6 @@ public class SubjectController {
 		return "subject/show_subject";
 	}
 
-	// UPDATE
 	@GetMapping("/update/{subjectId}")
 	public String updateSubject(Model m, @PathVariable long subjectId) {
 		Subject subject = this.subjectRepository.findOne(subjectId);
@@ -81,41 +76,28 @@ public class SubjectController {
 		this.subjectRepository.save(subject);
 		return "index";
 	}
-
-	// DELETE
+	
 	@GetMapping("/delete/{subjectId}")
+	public String deleteSubject(@PathVariable long subjectId, Model m) {
+		Subject subject = this.subjectRepository.findOne(subjectId);
+		m.addAttribute("subject", subject);
+		return "subject/confirmdelete_subject";
+	}
+
+	@PostMapping("/delete/{subjectId}")
 	public String deleteSubject(@PathVariable long subjectId) {
+		Subject subject = this.subjectRepository.findOne(subjectId);
+		if (subject.getSchool() != null || subject.getTeacher() != null || subject.getMark() != null || subject.getDivision() != null) {
+			return "errors/deleteException";
+		}
 		this.subjectRepository.delete(subjectId);
-		return "index";
+		return "redirect:/subject/all";
 	}
 	
-	// SHOW ALL
+	
 	@ModelAttribute("availableSubjects")
 	public List<Subject> getSubject() {
 		return this.subjectRepository.findAll();
 	}
-	
-	// MESSAGES INFO
-	@ModelAttribute("countAllReceivedMessages")
-	public Integer countAllReceivedMessages(Long receiverId) {
-		return this.messageRepository.findAllByReceiverId(getLoggedUser().getId()).size();
-	}
 
-	@ModelAttribute("countAllSendedMessages")
-	public Integer countAllSendedMessages(Long senderId) {
-		return this.messageRepository.findAllBySenderId(getLoggedUser().getId()).size();
-	}
-	
-	@ModelAttribute("countAllReceivedUnreadedMessages")
-	public Integer countAllReceivedUnreadedMessages(Long receiverId, Integer checked) {
-		return this.messageRepository.findAllByReceiverIdAndChecked(getLoggedUser().getId(), 0).size();
-	}
-	
-	private User getLoggedUser() {
-		Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-		String username = ((org.springframework.security.core.userdetails.User) principal).getUsername();
-		return this.userRepository.findOneByUsername(username);
-	}
-	
-	
 }
