+++ src/main/java/pl/schoolmanager/controller/TeacherController.java
@@ -2,12 +2,15 @@ package pl.schoolmanager.controller;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.ListIterator;
 
+import javax.persistence.PersistenceException;
 import javax.servlet.http.HttpSession;
+import javax.validation.ConstraintViolationException;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.orm.jpa.JpaSystemException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -18,16 +21,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import pl.schoolmanager.bean.SessionManager;
-import pl.schoolmanager.entity.School;
-import pl.schoolmanager.entity.Subject;
-import pl.schoolmanager.entity.Teacher;
-import pl.schoolmanager.entity.User;
-import pl.schoolmanager.entity.UserRole;
+import pl.schoolmanager.bean.role.Role;
+import pl.schoolmanager.entity.*;
 import pl.schoolmanager.repository.MessageRepository;
 import pl.schoolmanager.repository.SchoolRepository;
 import pl.schoolmanager.repository.SubjectRepository;
 import pl.schoolmanager.repository.TeacherRepository;
-import pl.schoolmanager.repository.UserRepository;
 
 @Controller
 @RequestMapping("/teacher")
@@ -39,25 +38,20 @@ public class TeacherController {
 	@Autowired
 	private SubjectRepository subjectRepository;
 
-
-	@Autowired
-	private UserRepository userRepo;
 	@Autowired
 	private SchoolRepository schoolRepo;
-
-	
-
 
 	@Autowired
 	private MessageRepository messageRepository;
 
+	@Autowired
+	private SessionManager sessionManager;
 
 	@GetMapping("/all")
 	public String all(Model m) {
 		return "teacher/all_teachers";
 	}
 
-	// CREATE
 	@GetMapping("/create")
 	public String createTeacher(Model m) {
 		m.addAttribute("teacher", new Teacher());
@@ -76,6 +70,10 @@ public class TeacherController {
 	// Make new teacher automatically creating new user role
 	@GetMapping("/userNewTeacher")
 	public String newTeacherFromUser(Model m) {
+		HttpSession s = SessionManager.session();
+		s.setAttribute("thisSchoolAdmin", null);
+		s.setAttribute("thisTeacher", null);
+		s.setAttribute("thisStudent", null);
 		m.addAttribute("teacher", new Teacher());
 		return "teacher/user_new_teacher";
 	}
@@ -85,24 +83,29 @@ public class TeacherController {
 		if (bindingResult.hasErrors()) {
 			return "teacher/user_teacher";
 		}
-		User user = getLoggedUser();
+		User user = sessionManager.loggedUser();
 		UserRole userRole = new UserRole();
 		userRole.setUsername(user.getUsername());
 		userRole.setUserRole("ROLE_TEACHER");
 		userRole.setSchool(teacher.getSchool());
 		userRole.setUser(user);
+		userRole.setEnabled(false);
 		teacher.setUserRole(userRole);
 		this.teacherRepository.save(teacher);
 		HttpSession s = SessionManager.session();
 		s.setAttribute("thisTeacher", teacher);
 		s.setAttribute("thisSchool", teacher.getSchool());
-		return "redirect:/teacherView/all";
+		return "user/user_activation_info";
 	}
-	
+
 	// Managing existing teacher role
 	@GetMapping("/userTeacher")
 	public String TeacherFromUser(Model m) {
 		m.addAttribute("teacher", new Teacher());
+		HttpSession s = SessionManager.session();
+		s.setAttribute("thisSchoolAdmin", null);
+		s.setAttribute("thisTeacher", null);
+		s.setAttribute("thisStudent", null);
 		return "teacher/user_teacher";
 	}
 
@@ -111,12 +114,13 @@ public class TeacherController {
 		if (bindingResult.hasErrors()) {
 			return "teacher/user_teacher";
 		}
-		User user = getLoggedUser();
+		User user = sessionManager.loggedUser();
 		Teacher thisTeacher = null;
 		UserRole thisUserRole = null;
 		List<UserRole> userRoles = user.getUserRoles();
 		for (UserRole userRole : userRoles) {
-			if (userRole.getUserRole().equals("ROLE_TEACHER") && userRole.getSchool().getName().equals(teacher.getSchool().getName())) {
+			if (userRole.getUserRole().equals("ROLE_TEACHER")
+					&& userRole.getSchool().getName().equals(teacher.getSchool().getName())) {
 				thisUserRole = userRole;
 			}
 		}
@@ -129,10 +133,9 @@ public class TeacherController {
 		HttpSession s = SessionManager.session();
 		s.setAttribute("thisTeacher", thisTeacher);
 		s.setAttribute("thisSchool", teacher.getSchool());
-		return "redirect:/teacherView/all";
+		return "redirect:/teacherView/";
 	}
 
-	// READ
 	@GetMapping("/view/{teacherId}")
 	public String viewTeacher(Model m, @PathVariable long teacherId) {
 		Teacher teacher = this.teacherRepository.findOne(teacherId);
@@ -140,7 +143,6 @@ public class TeacherController {
 		return "teacher/show_teacher";
 	}
 
-	// UPDATE
 	@GetMapping("/update/{teacherId}")
 	public String updateTeacher(Model m, @PathVariable long teacherId) {
 		Teacher teacher = this.teacherRepository.findOne(teacherId);
@@ -159,20 +161,36 @@ public class TeacherController {
 		return "index";
 	}
 
-	// DELETE
 	@GetMapping("/delete/{teacherId}")
-	public String deleteTeacher(@PathVariable long teacherId) {
-		this.teacherRepository.delete(teacherId);
-		return "index";
+	public String deleteTeacher(@PathVariable long teacherId, Model m) {
+		Teacher teacher = this.teacherRepository.findOne(teacherId);
+		m.addAttribute("teacher", teacher);
+		return "teacher/confirmdelete_teacher";
 	}
 
-	// SHOW ALL
+	@PostMapping("/delete/{teacherId}")
+	public String deleteSchool(@PathVariable long teacherId) {
+		Teacher teacher = this.teacherRepository.findOne(teacherId);
+		if (teacher.getSubject() != null || teacher.getSchool() != null) {
+			return "errors/deleteException";
+		}
+		this.teacherRepository.delete(teacherId);
+		return "redirect:/teacher/all";
+	}
+
 	@ModelAttribute("availableTeachers")
 	public List<Teacher> getTeachers() {
 		return this.teacherRepository.findAll();
 	}
 
-	// ADD SUBJECT TO TEACHER
+	// SHOW ALL FROM SCHOOL
+	@ModelAttribute("schoolTeachers")
+	public List<Teacher> getSchoolTeachers() {
+		HttpSession s = SessionManager.session();
+		School school = (School) s.getAttribute("thisSchool");
+		return this.teacherRepository.findAllBySchool(school);
+	}
+
 	@GetMapping("/addSubject/{teacherId}")
 	public String addSubject(Model m, @PathVariable long teacherId) {
 		Teacher teacher = this.teacherRepository.findOne(teacherId);
@@ -193,12 +211,18 @@ public class TeacherController {
 		return "redirect:/teacher/addSubject/{teacherId}";
 	}
 
-
-	// Additional methods
-	private User getLoggedUser() {
-		Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-		String username = ((org.springframework.security.core.userdetails.User) principal).getUsername();
-		return this.userRepo.findOneByUsername(username);
+	@GetMapping("removeSubject/{teacherId}/{subjectId}")
+	public String removeSubject(@PathVariable long teacherId, @PathVariable long subjectId) {
+		Teacher teacher = this.teacherRepository.findOne(teacherId);
+		Subject subject = this.subjectRepository.findOne(subjectId);
+		ListIterator<Teacher> teach = subject.getTeacher().listIterator();
+		while (teach.hasNext()) {
+			if (teach.next().getId() == teacher.getId()) {
+				teach.remove();
+			}
+		}
+		this.subjectRepository.save(subject);
+		return "redirect:/teacher/addSubject/{teacherId}";
 	}
 
 	@ModelAttribute("availableSchools")
@@ -206,34 +230,31 @@ public class TeacherController {
 		List<School> availableSchools = this.schoolRepo.findAll();
 		return availableSchools;
 	}
-	// MESSAGES INFO
-	@ModelAttribute("countAllReceivedMessages")
-	public Integer countAllReceivedMessages(Long receiverId) {
-		return this.messageRepository.findAllByReceiverId(getLoggedUser().getId()).size();
-	}
-
-	@ModelAttribute("countAllSendedMessages")
-	public Integer countAllSendedMessages(Long senderId) {
-		return this.messageRepository.findAllBySenderId(getLoggedUser().getId()).size();
-	}
-	
-	@ModelAttribute("countAllReceivedUnreadedMessages")
-	public Integer countAllReceivedUnreadedMessages(Long receiverId, Integer checked) {
-		return this.messageRepository.findAllByReceiverIdAndChecked(getLoggedUser().getId(), 0).size();
-	}
-	
 
 	@ModelAttribute("userSchools")
 	public List<School> userSchools() {
-		User user = getLoggedUser();
+		User user = sessionManager.loggedUser();
 		List<School> schools = new ArrayList<>();
 		List<UserRole> roles = user.getUserRoles();
 		for (UserRole userRole : roles) {
-			if (userRole.getUserRole().equals("ROLE_TEACHER")) {
+			if (Role.TEACHER.isEqualTo(userRole.getUserRole())) {
 				schools.add(userRole.getSchool());
 			}
 		}
 		return schools;
 	}
 
+	@ModelAttribute("userEnabledSchools")
+	public List<School> userEnabledSchools() {
+		User user = sessionManager.loggedUser();
+		List<School> schools = new ArrayList<>();
+		List<UserRole> roles = user.getUserRoles();
+		for (UserRole userRole : roles) {
+			if (userRole.getUserRole().equals("ROLE_TEACHER") && userRole.isEnabled() == true) {
+				schools.add(userRole.getSchool());
+			}
+		}
+		return schools;
+	}
+	
 }
